Update ProjectCard
To improve the user experience (UX), I added an <a> tag to the project image. The tag defaults to the live URL of the website. If a live URL does not exist, it replaces it with a GitHub URL instead. Finally, if neither a live URL nor GitHub is presented, it does not add a link. This change allows users to easily view a project without having to think much about where to go.